move mc_crypt.x to enclave/
